Reviewed some code; addressed some TODOs. - --session --display was iterating incorrectly; if there were many sessions for a given IPv4 "local" transport address, it would loop indefinitely. - Bettered prefix handling for extreme cases. Prefix 0.0.0.0/0 was triggering some "1 << 32"s, which is an undefined operation in C. - Updated unit tests.
I still need to address a kernel panic my tester has and which I'm being unable to reproduce. It happens by modprobing jool_siit after removing jool.
Also, the new unit test suggests more revision is in order: http://stackoverflow.com/questions/29551002/copying-a-32-bit-integer-to-a-64-bit-integer-is-undefined

+++ include/nat64/mod/common/address.h
@@ -38,7 +38,7 @@ bool prefix6_contains(const struct ipv6_prefix *prefix, const struct in6_addr *a
 
 bool prefix4_intersects(const struct ipv4_prefix *p1, const struct ipv4_prefix *p2);
 
-unsigned int prefix4_get_addr_count(struct ipv4_prefix *prefix);
+__u64 prefix4_get_addr_count(struct ipv4_prefix *prefix);
 
 int prefix4_validate(struct ipv4_prefix *prefix);
 int prefix6_validate(struct ipv6_prefix *prefix);
